Finally actually fix motion
It's super messy but it fixes the underlying issue—the tile transition
was calculated such in a point like fashion in so would sometimes the
snake would appear as if it had gone to another tile when changing
directions. Now the tile boundary is at the leading edge of the snake
head and we base direction switching by looking ahead to the next tile
than seeing if the current tile is different from the previous (which
meant it was too late!). My thinking was too wooly and lazy to begin
with which is what got me into this mess to begin with.

The solution is a terrific hack in parts (SnakeBody is a tire fire)
but could probably be tidied up if I could be bothered.